+++ tests/unit/resources/storage/test_storage_volume_templates.py
@@ -102,8 +102,8 @@ class StorageVolumeTemplatesTest(unittest.TestCase):
 
     @mock.patch.object(ResourceClient, 'get')
     def test_get_reachable_volume_templates_called_once(self, mock_get):
-        uri = '/rest/storage-volume-templates/reachable-volume-templates?networks=\'/rest/fake\'&start=0&count=-1'
-        self._storage_volume_templates.get_reachable_volume_templates(networks=['/rest/fake'])
+        uri = '/rest/storage-volume-templates/reachable-volume-templates?networks=/rest/fake&privateAllowedOnly=False&start=0&count=-1'
+        self._storage_volume_templates.get_reachable_volume_templates(networks='/rest/fake')
         mock_get.assert_called_once_with(uri)
 
     @mock.patch.object(ResourceClient, 'get')